refactor: v1.2.6 整理工程模式代码，优化按键与UI逻辑
本次提交包含多项改进：
1. 重构工程模式后台任务逻辑，将参数调整逻辑移出UI层
2. 修正按键左右切Tab的映射关系
3. 优化DCAC过载检测逻辑，移除冗余的bDcac_GetOverLoadState接口
4. 调整调试开关、BMS充电电压默认值，缩短部分任务周期
5. 修复错误码宏定义的位数溢出问题
6. 简化升级流程中的临界区保护逻辑
7. 补充系统保护检测的BMS额外告警项
8. 调整UI编辑模式的边框样式与交互逻辑
9. 修复部分注释与文档描述不一致的问题

+++ APP/Application/main.h
@@ -88,11 +88,11 @@ OF SUCH DAMAGE.
 
 #define     	BIT_GET(obj, XX_CODE)        			((obj >> XX_CODE) & (0x01))
 
-#define     	ERR_SET(obj, ERR_CODE)         			BIT_SET(obj, ((u16)0x1 << ERR_CODE))
-#define     	ERR_CLR(obj, ERR_CODE)         			BIT_CLR(obj, ((u16)0x1 << ERR_CODE))
+#define     	ERR_SET(obj, ERR_CODE)         			BIT_SET(obj, ((u64)0x1 << ERR_CODE))
+#define     	ERR_CLR(obj, ERR_CODE)         			BIT_CLR(obj, ((u64)0x1 << ERR_CODE))
 
-#define     	STAT_SET(obj, STAT_CODE)        		BIT_SET(obj, ((u16)0x1 << STAT_CODE))
-#define     	STAT_CLR(obj, STAT_CODE)        		BIT_CLR(obj, ((u16)0x1 << STAT_CODE))
+#define     	STAT_SET(obj, STAT_CODE)        		BIT_SET(obj, ((u64)0x1 << STAT_CODE))
+#define     	STAT_CLR(obj, STAT_CODE)        		BIT_CLR(obj, ((u64)0x1 << STAT_CODE))
